tests/wpt/tests/content-security-policy/unsafe-hashes/javascript_src_allowed-window_open.html

<!DOCTYPE HTML>
<html>

<head>
    <meta http-equiv="Content-Security-Policy" content="script-src 'unsafe-hashes' 'nonce-abc'
    'sha256-IIiAJ8UuliU8o1qAv6CV4P3R8DeTf/v3MrsCwXW171Y=';">
    <!--
    'sha256-IIiAJ8UuliU8o1qAv6CV4P3R8DeTf/v3MrsCwXW171Y=' ==> 'javascript:opener.postMessage('pass', '*')'
    -->
    <script src='/resources/testharness.js' nonce='abc'></script>
    <script src='/resources/testharnessreport.js' nonce='abc'></script>
</head>

<body>
    <div id='log'></div>
    <script nonce='abc'>
        var t1 = async_test("Test that the javascript: src is allowed to run");

        window.onmessage = t1.step_func_done(function(e) {
          assert_equals(e.data, "pass");
        });

        window.addEventListener('securitypolicyviolation', t1.unreached_func("Should have not raised any event"));

        window.open("javascript:opener.postMessage('pass', '*')");
    </script>
</body>

</html>
